AD654ok之后 备份

+++ Time_test.h
@@ -10,8 +10,8 @@
 #define START_TIME  time=0;RCC_APB1PeriphClockCmd(RCC_APB1Periph_TIM2 , ENABLE);TIM_Cmd(TIM2, ENABLE)
 #define STOP_TIME  TIM_Cmd(TIM2, DISABLE);RCC_APB1PeriphClockCmd(RCC_APB1Periph_TIM2 , DISABLE)
 
-extern u16 EncoderData;
-extern u16 temp;
+extern u32 EncoderData;
+extern u32 temp;
 extern int	ValvePosValue ;		//三位数，最后一位是小数位 
 //前一次的编码器值
 extern int	ValvePosValueLast ;		//三位数，最后一位是小数位
